requirements.txt
Django==1.3
distribute==0.6.14
django-breadcrumbs==1.1.0
django-debug-toolbar==0.8.4
django-mailer==0.1.0
django-notification==0.1.5
django-pagination==1.0.7
django-registration==0.7
django-treemenus==0.8.7-pre
html5lib==0.90
pisa==3.0.33
psycopg2==2.4
reportlab==2.5
simplejson==2.1.3
wsgiref==0.1.2



